requirement.txt
Pillow
tk
ttkthemes
PyQt5
mysql-connector
os
opencv-python
numpy